Revised several textures, removed unused features, worked on rituals

+++ src/main/java/com/xX_deadbush_Xx/witchcraftmod/api/ritual/IContinuousRitual.java
@@ -1,7 +1,7 @@
 package com.xX_deadbush_Xx.witchcraftmod.api.ritual;
 
 public interface IContinuousRitual {
-	int manaPerSecond();
+	double manaPerSecond();
 	
 	/**
 	 * Client and server side
